refactor(account-sorting): Moved the page select up a level
